# PERF: faster corrwith method for pearson and spearman correlation when other is a Series and axis = 0 (column-wise)