--- a/pandas/core/frame.py
+++ b/pandas/core/frame.py
@@ -9838,23 +9838,26 @@ NaN 12.3   33.0
         this = self._get_numeric_data()
 
         if isinstance(other, Series):
-            if method == "kendall" or axis == 1:
-                return this.apply(lambda x: other.corr(x, method=method), axis=axis)
-            else:
+            if axis == 0 and method in ["pearson", "spearman"]:
                 corrs = {}
                 ndf = self.values.transpose()
                 k = other.values
                 if method == "pearson":
                     for i, r in enumerate(ndf):
-                        nonnull_mask = (~np.isnan(r) & ~np.isnan(k))
-                        corrs[self.columns[i]] = np.corrcoef(r[nonnull_mask], k[nonnull_mask])[0, 1]
-                elif method == "spearman":
-                    for i, r in enumerate(ndf):
-                        nonnull_mask = (~np.isnan(r) & ~np.isnan(k))
-                        corrs[self.columns[i]] = np.corrcoef(r[nonnull_mask].argsort().argsort(), k[nonnull_mask].argsort().argsort())[0, 1]
+                        nonnull_mask = ~np.isnan(r) & ~np.isnan(k)
+                        corrs[self.columns[i]] = np.corrcoef(
+                            r[nonnull_mask], k[nonnull_mask]
+                        )[0, 1]
                 else:
-                    raise NotImplementedError(f"Invalid correlation method {method}")
+                    for i, r in enumerate(ndf):
+                        nonnull_mask = ~np.isnan(r) & ~np.isnan(k)
+                        corrs[self.columns[i]] = np.corrcoef(
+                            r[nonnull_mask].argsort().argsort(),
+                            k[nonnull_mask].argsort().argsort(),
+                        )[0, 1]
                 return Series(corrs)
+            else:
+                return this.apply(lambda x: other.corr(x, method=method), axis=axis)
 
         other = other._get_numeric_data()
         left, right = this.align(other, join="inner", copy=False)