--- a/pandas/core/frame.py
+++ b/pandas/core/frame.py
@@ -9840,7 +9840,7 @@ NaN 12.3   33.0
         if isinstance(other, Series):
             if axis == 0 and method in ["pearson", "spearman"]:
                 corrs = {}
-                ndf = self.values.transpose()
+                ndf = self.select_dtypes(include=np.number).values.transpose()
                 k = other.values
                 if method == "pearson":
                     for i, r in enumerate(ndf):