## Description of Changes

This PR modifies the `corrwith()` method defined in pandas/core/frame.py to be faster when all of the following (common) conditions are met.

- the `axis` argument is set to 0
- the `other` argument is a Pandas Series object
- the `method` argument is equal to "pearson" OR "spearman"

When all these are true, we replace the `apply()` on `self` with column-wise iteration on NumPy arrays and the correlation function against `other` is implemented in NumPy's `np.corrcoef()`. If not all of those conditions are met, the method falls back to its existing functionality with no performance change. The Spearman rank correlation is calculated via `np.corrcoef()` by taking the Pearson correlation coefficient of the vector element ranks with `argsort()`.

When `method` = "pearson", this change reduces execution time by 40 - 50% on small and large DataFrame objects.

When `method` = "spearman", this change reduces execution time by 70 - 80% on small and large DataFrame objects.

This change is still robust to null values on the left or right side despite the fact that `np.corrcoef()` is not robust to nulls, because it masks the null values with a boolean `&` across both objects when iterating the vectors. It returns accurate results.

## Testing Environment & Performance Assessment

My testing environment is as follows:

- Hardware: MacBook Pro (16-inc, 2021) with Apple M1 Max Chip and 64GB RAM
- OS: macOS Monterey v12.2
- Python: 3.8.12
- Pandas: v1.41 (Control, Unmodified) and v1.5.0.dev (Self-Built Test)
- NumPy: v1.22.2
- Installer/Package Manager: Anaconda3 (miniforge3 for Apple Silicon/arm64)

My test script was the following:

```
import timeit
import numpy as np
import pandas as pd

pearson_results = np.array([])
spearman_results = np.array([])
num_tests = 10
num_runs = 10

for k in range(num_tests):
    print(f'Running test {k + 1}...')
    data = np.random.uniform(-1, 1, size=(100, 50000))
    df = pd.DataFrame(data)
    df.columns = [str(c) for c in df.columns]
    pearson_time = timeit.timeit("df.corrwith(df['0'], axis=0, method='pearson')", "from __main__ import df", number=num_runs, timer=timeit.default_timer)
    spearman_time = timeit.timeit("df.corrwith(df['0'], axis=0, method='spearman')", "from __main__ import df", number=num_runs, timer=timeit.default_timer)
    pearson_results = np.append(pearson_results, pearson_time)
    spearman_results = np.append(spearman_results, spearman_time)

print(f'Average Pearson Correlation Time Across {num_tests} DataFrames with {num_runs} runs each: {pearson_results.mean() / num_runs}')
print(f'Average Spearman Correlation Time Across {num_tests} DataFrames {num_runs} runs each: {spearman_results.mean() / num_runs}')
```

First I created a fresh conda environment with the current stable version of Pandas (v1.41). For a DataFrame with 100 rows and 50,000 columns, using unmodified `corrwith()` with `method="pearson"` took approximately 3.0697 seconds on average across 10 runs each on 10 different randomly generated DataFrames. With `method="spearman"`, the same evaluation took approximately 8.5808 seconds on average.

Then I made my changes to my fork of Pandas main and built it in a new Conda environment, then ran the same script. For an equivalently sized DataFrame with random numerical data, the modified `corrwith()` with `method="pearson"` took approximately 1.5342 seconds. With `method="spearman"`, it took approximately 2.0556 seconds.