@fractionalhare 2022-02-28
I haven't used `asv` before, but I can generate some test results from doing so.