@fractionalhare 2022-02-28
@jreback I can't get asv to work on my machine, it has odd errors about hdf5 regardless of whether I run with conda or virtualenv. I'm not sure if this is related to me trying to run on it on an M1. If I try to run it with my existing environment I get the following error:

`No range spec may be specified if benchmarking in an existing environment`

However, you can see the results of just the relevant benchmark from the automated test suite at this line:
https://github.com/pandas-dev/pandas/runs/5354404445?check_suite_focus=true#step:6:8761