@jreback 2022-03-01
> Adding to discussion here: is it worth adding more tests of `corrwith` correctness with expected outputs for sample inputs, up to precision with something like `np.isclose()`?

yes