@jreback 2022-03-01
@fractionalhare show the asv result above (or at the very least show a timeit with before/after of the sample)