@fractionalhare 2022-03-02
Unclear why the ASV benchmark failed, a test in unrelated/unmodified code triggered an `exit code 1` in the benchmarks after I merged main into this branch, but the ASVs for `corrwith` remain clean: https://github.com/pandas-dev/pandas/runs/5392988518?check_suite_focus=true#step:6:9179

```
[ 39.33%] ··· stat_ops.Correlation.time_corrwith_cols                         ok
[ 39.33%] ··· ========== ==========
                method             
              ---------- ----------
               spearman   204±0ms  
               kendall    202±0ms  
               pearson    6.61±0ms 
              ========== ==========

[ 39.38%] ··· stat_ops.Correlation.time_corrwith_rows                         ok
[ 39.38%] ··· ========== ==========
                method             
              ---------- ----------
               spearman   384±0ms  
               kendall    366±0ms  
               pearson    6.84±0ms 
              ========== ==========
```

Tests were all green on this PR prior to me pulling main just now, so maybe this is ephemeral?